requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.11
django-jazzmin==2.6.1
django-livereload-server==0.5.1
django-widget-tweaks==1.5.0
pillow==10.2.0
python-decouple==3.8
sqlparse==0.4.4
tornado==6.4
typing_extensions==4.10.0
tzdata==2024.1
